refactor functional framework. introduce collector factory
