+++ internal/cmd/functional-benchmarker/runners/cluster/cluster_runner.go
@@ -129,7 +129,7 @@ func (r *ClusterRunner) Cleanup() {
 
 func (r *ClusterRunner) SampleCollector() *stats.Sample {
 	if result, err := oc.AdmTop(r.framework.Namespace, r.framework.Name).NoHeaders().ForContainers().Run(); err == nil {
-		log.V(3).Info("Sample Collector", "result", result)
+		log.V(3).Info("Sample collector", "result", result)
 		if !strings.Contains(result, "Error from server") {
 			for _, line := range strings.Split(result, "\n") {
 				fields := strings.Fields(line)
