+++ test/functional/outputs/forward_to_syslog_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/openshift/cluster-logging-operator/test"
 )
 
-var _ = Describe("[LogForwarding][Syslog] Functional tests", func() {
+var _ = Describe("[Functional][Outputs][Syslog] Functional tests", func() {
 
 	var (
 		framework *functional.FluentdFunctionalFramework
